dekita

+++ srcs/zoom.c
@@ -4,5 +4,5 @@ void	draw_zoom(t_point *point, double scale, int zoom)
 {
     point->x *= zoom;
     point->y *= zoom;
-    point->z *= (int)(scale * zoom);
+    point->z *= (int)(scale * (double)zoom);
 }
